Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,6 +2,6 @@
 void main() {
     int a,b,c;
     scanf("%d%d%d",&a,&b,&c);
-    avg=(a+b+c)/3;
+    avg=(float)(a+b+c)/3;
     printf("Average: %f",avg);
 }
